feat(awareness): ambient desktop + Initiative situational continuity for 2.8.5
Give Initiative and chat a cheap sense of time and place without burning
the hot path or stuffing the feed.

Ambient desktop awareness
- Capture a compact frontmost-app snapshot (a11y labels, not screenshots)
  on the Initiative schedule-tick background thread only.
- Serve the hot path from a single overwritten cache; never hit sidecar
  or walk the tree on chat/LLM assemble.
- Hash-dedup so unchanged desktops inject as a one-line skip; expire and
  delete the cache after 24h so stale context cannot rot prompts.
- Expose get_desktop_snapshot for deliberate pulls; system prompt notes
  that deeper tree tools still exist for targeting.

Initiative situational continuity
- Stamp cycle context with now, idle/cycle gaps, and a handoff peek so
  the model knows when the last real work was.
- After a long idle (≥2h), prefer resuming from a project handoff over
  unrelated board/work_scan noise.
- Surface developer_context captured_at and tighten Initiative settings
  copy so daily plans stay in Automations and Initiative stays one gated
  action.

Also lands the broader 2.8.5 dirty tree that this awareness work sits on:
workflow blueprint adherence and control policy, harness/steering and
coordination hardening, kanban/lifecycle and ticket contract fixes, CLI
model policy and provider preflight, Kokoro TTS descriptor work, and the
matching test coverage.

Changelog 2.8.5 gets a short reader note: awareness without slowing you
down.

+++ distr/core/agent/tools/input/sidecar_http.py
@@ -49,7 +49,7 @@ def is_sidecar_reachable(timeout: float = 2.0) -> bool:
     return sidecar_health(timeout=timeout) is not None
 
 
-def call_sidecar_tool(tool: str, params: dict, *, timeout: int = 120) -> dict:
+def call_sidecar_tool(tool: str, params: dict, *, timeout: float = 120) -> dict:
     """
     ``POST /tool/{tool}`` with JSON body. Raises ``RuntimeError`` on connection errors or HTTP failures.
     """
